feat: clean and organized code
